requirements.txt
openai
python-docx
python-dotenv
mutagen
tiktoken
google-generativeai
